requirements.txt
Flask
pillow
langchain
langchain-google-genai
google-generativeai
flask_cors